# openapi/dist/openapi.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Reel Fake Api
  description: A REST api for buying or renting a movie dvd (only for educational purposes)
  license:
    name: MIT License
    url: https://mit-license.org
  contact:
    name: Pratap Reddy
    email: pratapreddy15@gmail.com
servers:
  - url: http://reelfake.cloud/api
    description: Production server
  - url: http://127.0.0.1:8000/api
    description: Local server
security:
  - cookieAuth: []
paths:
  /actors:
    get:
      summary: Get the actor list with pagination support
      operationId: getActors
      tags:
        - actor
      parameters:
        - in: query
          name: page
          required: false
          description: |
            The page number (which starts with 1) to fetch the set of actors from.
            The actors are split into multiple pages with each page having a limited number of actors.
          schema:
            type: integer
            default: 1
        - in: query
          name: name
          required: false
          description: |
            The name of the actor. The api accepts ?name=keanu (for exact match) and ?name=%an% (for actor name having "an").
          schema:
            type: string
        - in: query
          name: birthday
          required: false
          description: |
            Filters the actors by the birth date (YYYY-MM-DD). This can be a single date or a range. For e.g., ?birthday=1980-01-01 or ?birthday=1980-01-01,1990-06-30. The query string ?birthday=1980-01-01 will query actors with the given birth date while ?birthday=1980-01-01,1990-06-30 will query actors born between 1980-01-01 and 1990-06-30.
          schema:
            type: string
            format: date
        - in: query
          name: deathday
          required: false
          description: |
            Filters the actors by the death date (YYYY-MM-DD). This can be a single date or a range. For e.g., ?deathday=1980-01-01 or ?deathday=1980-01-01,1990-06-30. The query string ?deathday=1980-01-01 will query actors with the given death date while ?deathday=1980-01-01,1990-06-30 will query actors died between 1980-01-01 and 1990-06-30.
          schema:
            type: string
            format: date
        - in: query
          name: place_of_birth
          required: false
          description: |
            The place of birth of the actor. The api accepts ?place_of_birth=jakarta (for exact match) and ?place_of_birth=%art% (for place of birth having "art").
          schema:
            type: string
        - in: query
          name: popularity
          required: false
          description: |
            The popularity index of the actor
          schema:
            type: integer
            format: float
      responses:
        '200':
          description: Successfully returned list of actor summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/actorSummary'
                  length:
                    type: integer
                  pagination:
                    $ref: '#/components/schemas/pagination'
        '400':
          $ref: '#/components/responses/badRequest'
        '500':
          $ref: '#/components/responses/internalServerError'
    post:
      summary: Add new actor
      operationId: addActor
      tags:
        - actor
      requestBody:
        description: Payload for the new actor
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tmdbId:
                  type: string
                imdbId:
                  type: string
                actorName:
                  type: string
                biography:
                  type: string
                birthday:
                  type: string
                  format: date
                deathday:
                  type: string
                  format: date
                placeOfBirth:
                  type: string
                popularity:
                  type: number
                  format: float
                profilePictureUrl:
                  type: string
                  format: uri
      responses:
        '201':
          description: Successfully added new actor
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  imdbId:
                    type: string
                  actorName:
                    type: string
                  biography:
                    type: string
                  birthday:
                    type: string
                    format: date
                  deathday:
                    type: string
                    format: date
                  placeOfBirth:
                    type: string
                  popularity:
                    type: number
                    format: float
                  profilePictureUrl:
                    type: string
                    format: uri
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '500':
          $ref: '#/components/responses/internalServerError'
  /actors/{id}:
    get:
      summary: Get actor detail using the id
      operationId: getActorById
      tags:
        - actor
      parameters:
        - in: path
          name: id
          required: true
          description: The id of the actor that is requested
          schema:
            type: integer
        - in: query
          name: includeMovies
          required: false
          description: |
            Whether or not to include the movies for the actor. Accepted values are true, false, yes, no, 0 and 1.
            If includeMovies value is truthy then the actor object with the movies are returned.
            If this is not present in query then it will default to false. If any value other than the accepted ones are
            given, then 400 is returned.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Successfully returned actorDetail object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/actorDetail'
        '400':
          $ref: '#/components/responses/badRequest'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
    post:
      summary: Add the actor to the movie
      operationId: addToMovie
      tags:
        - actor
      parameters:
        - in: path
          name: id
          required: true
          description: The id of the actor
          schema:
            type: integer
      requestBody:
        description: Payload for the actor to add to the movie
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                movieId:
                  type: integer
                  format: number
                characterName:
                  type: string
                castOrder:
                  type: integer
                  format: number
      responses:
        '201':
          description: Successfully added the actor the movie
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  actorId:
                    type: integer
                  movieId:
                    type: integer
                  characterName:
                    type: string
                  castOrder:
                    type: integer
                    format: number
                required:
                  - id
                  - actorId
                  - movieId
                  - characterName
                  - castOrder
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
    put:
      summary: Update actor
      operationId: updateActor
      tags:
        - actor
      parameters:
        - in: path
          name: id
          required: true
          description: Id of the actor
          schema:
            type: integer
      requestBody:
        description: Payload of the actor to update
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tmdbId:
                  type: integer
                  format: number
                imdbId:
                  type: integer
                  format: number
                actorName:
                  type: string
                biography:
                  type: string
                birthday:
                  type: string
                  format: date
                deathday:
                  type: string
                  format: date
                placeOfBirth:
                  type: string
                popularity:
                  type: string
                profilePictureUrl:
                  type: string
                  format: uri
      responses:
        '204':
          description: Successfully updated actor
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
    delete:
      summary: |
        Delete the actor.
        Note: This operation will delete movie actors for the movie as well.
      operationId: deleteActor
      tags:
        - actor
      parameters:
        - in: path
          name: id
          required: true
          description: The id of the actor
          schema:
            type: integer
      responses:
        '204':
          description: Successfully deleted the movie
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
...
  /auth/me:
    get:
      summary: Get user detail for the logged in user
      operationId: getUserProfile
      security:
        - cookieAuth: []
      tags:
        - authentication
      responses:
        '200':
          description: Successfully returned the logged in user detail
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/customerDetail'
                  - $ref: '#/components/schemas/staffItem'
        '400':
          $ref: '#/components/responses/badRequest'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      name: auth_token
      in: cookie
  schemas:
    actorSummary:
      type: object
      properties:
        id:
          type: integer
        imdbId:
          type: integer
        actorName:
          type: string
        birthday:
          type: string
          format: date
        deathday:
          type: string
          format: date
        placeOfBirth:
          type: string
        popularity:
          type: number
          format: float
        profilePictureUrl:
          type: string
          format: uri
    pagination:
      type: object
      properties:
        pageNumber:
          type: integer
          description: Current page number
        totalPages:
          type: integer
          description: Total pages available for pagination
        totalItems:
          type: integer
          description: Total items across all the pages
        itemsPerPage:
          type: integer
          description: Maximum number of items that the page can have
        next:
          type: string
          description: Link to the next page
        prev:
          type: string
          description: Link to the previous page
        first:
          type: string
          description: Link to the first page
        last:
          type: string
          description: Link to the last page
    actorDetail:
      type: object
      properties:
        id:
          type: integer
        imdbId:
          type: string
        actorName:
          type: string
        biography:
          type: string
        birthday:
          type: string
          format: date
        deathday:
          type: string
          format: date
        placeOfBirth:
          type: string
        popularity:
          type: number
          format: float
        profilePictureUrl:
          type: string
          format: uri
        movies:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              title:
                type: string
              releaseDate:
                type: string
                format: date
              genres:
                type: array
                items:
                  type: string
              ratingAverage:
                type: number
                format: float
              ratingCount:
                type: integer
              posterUrl:
                type: string
              cast:
                type: object
                properties:
                  characterName:
                    type: string
                  castOrder:
                    type: integer
                    format: number
    address:
      type: object
      properties:
        id:
          type: integer
          format: number
          readOnly: true
        addressLine:
          type: string
        cityName:
          type: string
        stateName:
          type: string
        country:
          type: string
        postalCode:
          type: string
    customerDetail:
      type: object
      properties:
        id:
          type: integer
          format: number
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        active:
          type: boolean
        phoneNumber:
          type: string
        avatar:
          type: string
          format: uri
        registeredOn:
          type: string
          format: date
        preferredStore:
          type: object
          properties:
            id:
              type: integer
              format: number
            phoneNumber:
              type: string
            storeManager:
              type: object
              properties:
                id:
                  type: string
                  format: number
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
                active:
                  type: boolean
                phoneNumber:
                  type: string
                avatar:
                  type: string
                  format: uri
                address:
                  $ref: '#/components/schemas/address'
        address:
          $ref: '#/components/schemas/address'
    staffItem:
      type: object
      properties:
        id:
          type: string
          format: number
          readOnly: true
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        isStoreManager:
          type: boolean
        active:
          type: boolean
        phoneNumber:
          type: string
        avatar:
          type: string
          format: uri
        address:
          $ref: '#/components/schemas/address'
...
  responses:
    badRequest:
      description: The request does not have the data that api needed
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    internalServerError:
      description: Server encountered unexpeted error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    unauthorized:
      description: Api key is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
    notFound:
      description: Requested resource does not exist
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    forbidden:
      description: User do not have permission to perform the operation
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string